fix: Corrige exibição de equipamentos no frontend
- Fix: Altera signal_dbm de Optional[int] para Optional[float] no schema Pydantic
  * Resolve ResponseValidationError que impedia listagem de equipamentos
  * Banco de dados armazena valores decimais (-3.13, 24.75, etc)

- Fix: Corrige porta do proxy Vite de 8080 para 8000
  * Frontend agora se comunica corretamente com o backend
  * Resolve erro de conexão no login e carregamento de dados

Resultado: 41 equipamentos agora aparecem corretamente na aba Equipamentos

+++ backend/app/schemas.py
@@ -41,7 +41,7 @@ class EquipmentBase(BaseModel):
     # Wireless
     brand: str = "generic"
     equipment_type: str = "station"  # "transmitter" or "station"
-    signal_dbm: Optional[int] = None
+    signal_dbm: Optional[float] = None
     ccq: Optional[int] = None
     connected_clients: Optional[int] = None
     whatsapp_groups: Optional[List[str]] = []
@@ -78,7 +78,7 @@ class EquipmentUpdate(BaseModel):
     api_port: Optional[int] = None
     brand: Optional[str] = None
     equipment_type: Optional[str] = None
-    signal_dbm: Optional[int] = None
+    signal_dbm: Optional[float] = None
     ccq: Optional[int] = None
     connected_clients: Optional[int] = None
     whatsapp_groups: Optional[List[str]] = None
